#109 [CHORE] 클라이언트 wrapper 분리
